pyproject.toml
[tool.poetry]
name = "alopekis"
version = "0.1.0"
description = "DataCite Data Files Generator"
authors = ["Mike Bennett <mbennett@datacite.org>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
opensearch-dsl = "^2.1.0"
ujson = "^5.10.0"
python-dotenv = "^1.0.1"
pyhumps = "^3.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
